modified the library to be easier to include outside of Arduino.
